pyproject.toml
[project]
name = "osm-python-readers-benchmark"
version = "0.1.0"
description = "Benchamr code used to analyse code performance of Python-based OSM readers."
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "duckdb==1.3.0",
    "esy-osm-shape==0.3",
    "gdal==3.11.0",
    "osmium==4.0.2",
    "osmnx==2.0.4",
    "pooch==1.8.2",
    "psutil==7.0.0",
    "pydriosm==2.2.0",
    "pyhelpers==2.0.0",
    "pyrosm==0.6.2",
    "quackosm[cli]==0.14.1",
]
